Move MeshHitShape3D to bits project and make it unique per mesh instance.
